refactor babylon creation objects and game errors, fixed ws ending
